+++ app/src/main/java/com/example/tasky/agenda_details/presentation/utils/DateTimeHelper.kt
@@ -1,0 +1,36 @@
+package com.example.tasky.agenda_details.presentation.utils
+
+import java.time.Duration
+import java.time.Instant
+import java.time.LocalDate
+import java.time.LocalTime
+import java.time.ZoneId
+import java.time.ZonedDateTime
+
+object DateTimeHelper {
+    fun getLocalDateFromEpoch(
+        epochSeconds: Long,
+        zoneId: ZoneId = ZoneId.systemDefault()
+    ): LocalDate {
+        val instant = Instant.ofEpochSecond(epochSeconds)
+        val zonedDateTime = ZonedDateTime.ofInstant(instant, zoneId)
+        return zonedDateTime.toLocalDate()
+    }
+
+    fun getLocalTimeFromEpoch(
+        epochSeconds: Long,
+        zoneId: ZoneId = ZoneId.systemDefault()
+    ): LocalTime {
+        val instant = Instant.ofEpochSecond(epochSeconds)
+        val zonedDateTime = ZonedDateTime.ofInstant(instant, zoneId)
+        return zonedDateTime.toLocalTime()
+    }
+
+    fun calculateTimeDifferenceInSeconds(startTime: Long, endTime: Long): Long {
+        val startTimeInSeconds = Instant.ofEpochSecond(startTime)
+        val endTimeInSeconds = Instant.ofEpochSecond(endTime)
+
+        val duration = Duration.between(startTimeInSeconds, endTimeInSeconds)
+        return duration.seconds
+    }
+}
